chore(suite-desktop): refactor (splitting into modules and libs)
