+++ packages/suite-desktop/package.json
@@ -5,10 +5,12 @@
     "private": true,
     "author": "SatoshiLabs <info@satoshilabs.com>",
     "homepage": "https://trezor.io/",
-    "main": "dist/electron.js",
+    "main": "dist/app.js",
     "scripts": {
-        "dev": "yarn copy-files && yarn build:electron && node scripts/dev.js",
+        "dev": "yarn dev:prepare && cross-env LAUNCH_ELECTRON=true node scripts/dev.js",
         "dev:run": "electron .",
+        "dev:prepare": "yarn copy-files && yarn build:lib",
+        "dev:local": "node scripts/dev.js",
         "clean": "rimraf ./build-electron && rimraf ./build && rimraf ./dist",
         "copy-files": "yarn workspace @trezor/suite-data copy-static-files",
         "build:electron": "rimraf ./dist && node scripts/build.js",
@@ -20,8 +22,12 @@
         "publish:linux": "yarn build:linux && electron-builder --publish always --linux",
         "publish:mac": "yarn build:mac && electron-builder --publish always --mac",
         "publish:win": "yarn build:win && electron-builder --publish always --win",
-        "lint": "eslint '**/*{.ts,.tsx}'",
-        "type-check": "tsc --project tsconfig.json",
+        "lint": "yarn lint:app && yarn lint:ui",
+        "lint:app": "eslint 'src-electron/**/*.ts' --parser-options=project:./src-electron/tsconfig.json",
+        "lint:ui": "eslint 'src/**/*{.ts,.tsx}'",
+        "type-check": "yarn type-check:app && yarn type-check:ui",
+        "type-check:app": "tsc --project src-electron/tsconfig.json",
+        "type-check:ui": "tsc --project tsconfig.json",
         "test:unit": "jest"
     },
     "build": {
