+++ packages/suite-desktop/scripts/build.js
@@ -5,8 +5,10 @@ const pkg = require('../package.json');
 const electronSource = path.join(__dirname, '..', 'src-electron');
 const isDev = process.env.NODE_ENV !== 'production';
 
+console.log('[Electron Build] Starting...');
+const hrstart = process.hrtime();
 build({
-    entryPoints: ['electron', 'preload'].map(f => path.join(electronSource, `${f}.ts`)),
+    entryPoints: ['app', 'preload'].map(f => path.join(electronSource, `${f}.ts`)),
     platform: 'node',
     bundle: true,
     target: 'node12.18.2', // Electron 11
@@ -18,4 +20,12 @@ build({
     sourcemap: isDev,
     minify: !isDev,
     outdir: path.join(__dirname, '..', 'dist'),
-}).catch(() => process.exit(1));
+})
+    .then(() => {
+        const hrend = process.hrtime(hrstart);
+        console.log(
+            '[Electron Build] Finished in %dms',
+            (hrend[1] / 1000000 + hrend[0] * 1000).toFixed(1),
+        );
+    })
+    .catch(() => process.exit(1));
