+++ packages/suite-desktop/scripts/dev.js
@@ -1,6 +1,7 @@
 const path = require('path');
 const { exec } = require('child_process');
 
+const launchElectron = process.env.LAUNCH_ELECTRON === 'true';
 const projectRoot = path.join(__dirname, '..');
 
 const { createServer } = require('http');
@@ -12,6 +13,10 @@ const next = require('next')({
 next.prepare().then(() => {
     const requestHandler = next.getRequestHandler();
     const server = createServer(requestHandler).listen(8000, () => {
+        if (!launchElectron) {
+            return;
+        }
+
         const electron = exec('yarn run dev:run', {
             cwd: projectRoot,
         });
