+++ packages/suite-desktop/src-electron/libs/menu.ts
