+++ packages/suite-desktop/src-electron/libs/processes/BaseProcess.ts
