+++ packages/suite-desktop/src-electron/libs/utils.ts
