+++ packages/suite-desktop/src-electron/modules/dev-tools.ts
@@ -1,0 +1,10 @@
+/**
+ * Enable development tools
+ */
+import { BrowserWindow } from 'electron';
+
+const init = (window: BrowserWindow) => {
+    window.webContents.openDevTools();
+};
+
+export default init;
