+++ packages/suite-desktop/src-electron/modules/menu.ts
@@ -1,0 +1,10 @@
+import { BrowserWindow, Menu } from 'electron';
+
+import { buildMainMenu } from '@lib/menu';
+
+const init = (window: BrowserWindow) => {
+    Menu.setApplicationMenu(buildMainMenu());
+    window.setMenuBarVisibility(false);
+};
+
+export default init;
