--- packages/suite-desktop/src-electron/store.ts
@@ -1,73 +1,0 @@
-import Store from 'electron-store';
-import { BrowserWindow, screen } from 'electron';
-
-// creates config.json inside appData folder https://electronjs.org/docs/api/app#appgetpathname
-const store = new Store();
-
-export const MIN_WIDTH = 770;
-export const MIN_HEIGHT = 700;
-export const MAX_WIDTH = 1920;
-export const MAX_HEIGHT = 1080;
-export const BUFFER = 0.2;
-
-const getInitialWindowSize = () => {
-    const { bounds } = screen.getPrimaryDisplay();
-    const buffer = {
-        width: bounds.width * BUFFER,
-        height: bounds.height * BUFFER,
-    };
-
-    let width = bounds.width - buffer.width;
-    if (width <= MIN_WIDTH) {
-        width = MIN_WIDTH;
-    } else if (width >= MAX_WIDTH) {
-        width = MAX_WIDTH;
-    }
-
-    let height = bounds.height - buffer.height;
-    if (height <= MIN_HEIGHT) {
-        height = MIN_HEIGHT;
-    } else if (height >= MAX_HEIGHT) {
-        height = MAX_HEIGHT;
-    }
-
-    return {
-        width,
-        height,
-    };
-};
-
-//
-
-export const getWinBounds = () => {
-    const { width, height } = getInitialWindowSize();
-    const winBounds = store.get('winBounds', { width, height });
-    return winBounds;
-};
-
-export const setWinBounds = (window: BrowserWindow) => {
-    const bounds = window.getBounds();
-    // save only non zero dimensions
-    if (bounds.width > 0 && bounds.height > 0) {
-        store.set('winBounds', bounds);
-    }
-};
-
-export const getUpdateSettings = () => {
-    return store.get('updateSettings', { skipVersion: '' });
-};
-
-export const setUpdateSettings = (updateSettings: UpdateSettings) => {
-    store.set('updateSettings', updateSettings);
-};
-
-export const getTorSettings = () => {
-    return store.get('torSettings', {
-        running: false,
-        address: '127.0.0.1:9050',
-    });
-};
-
-export const setTorSettings = (torSettings: TorSettings) => {
-    store.set('torSettings', torSettings);
-};
