+++ packages/suite-desktop/src-electron/tsconfig.json
@@ -1,37 +1,38 @@
 {
-  "compilerOptions": {
-    "watch": false,
-    "preserveWatchOutput": true,
-    "outDir": "../dist",
-    "noEmit": false,
-    "jsx": "preserve",
-    "target": "esnext",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "lib": [
-      "dom",
-      "es2019",
-      "esnext.asynciterable"
+    "compilerOptions": {
+        "watch": false,
+        "preserveWatchOutput": true,
+        "outDir": "../dist",
+        "noEmit": false,
+        "jsx": "preserve",
+        "target": "esnext",
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "lib": [
+            "dom",
+            "es2019",
+            "esnext.asynciterable"
+        ],
+        "forceConsistentCasingInFileNames": true,
+        "esModuleInterop": true,
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "skipLibCheck": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "noImplicitAny": true,
+        "baseUrl": ".",
+        "paths": {
+            "@lib/*": ["./libs/*"],
+            "@module/*": ["./modules/*"],
+            "@suite/*": ["../../suite/src/*"]
+        }
+    },
+    "include": [
+        "**/*.ts"
     ],
-    "forceConsistentCasingInFileNames": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "skipLibCheck": true,
-    "baseUrl": ".",
-    "paths": {
-      "@suite/*": ["../../suite/src/*"]
-    }
-  },
-  "include": [
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude": [
-    "**/*.js",
-    "node_modules",
-    ".next/**",
-    "dist/**"
-  ]
+    "exclude": [
+        "**/*.js",
+    ]
 }
