+++ packages/suite-desktop/tsconfig.json
@@ -20,6 +20,7 @@
         "**/*.tsx"
     ],
     "exclude": [
-        "**/*.js"
+        "**/*.js",
+        "src-electron/**/*.ts"
     ]
 }
